Add references to Event entity from EventTerm
EventTerm entity is now aware of Event, and is referencing it via it's
FK which points to Event.id field.

This change is accompanied by a migration file 'Version20170810082330'.

An Event should display all of the Terms that were at some point
associated with it, so I've added a collection of EventTerms to the
Event entity.