Add constructor to Event and clean up some setters
Added all properties to constructor, and decided to remove some of
the setters, because some Event information should not be changable:
- id
- name
- owner

Updated the CreateEventHandler to deal with the change in entity.

This will break the Entity editing, because 'edit' uses the same form
that's being used for creating a new Event.